+++ test.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 int main() {
-	float x = 12560 * 183;
+	long x = 12560 * 183;
 	cout << x << endl;
    
     return 0;
